requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==43.0.0
Django==5.0.7
django-active-link==0.2.2
django-admin-volt==1.0.10
django-allauth==0.63.6
django-ash==0.0.7
django-ckeditor==6.7.1
django-js-asset==2.2.0
django-mathfilters==1.0.0
django-phonenumber-field==8.0.0
idna==3.7
jwt==1.3.1
phonenumbers==8.13.40
pillow==10.4.0
psycopg==3.2.1
psycopg2==2.9.9
pycparser==2.22
pycryptodome==3.20.0
PyJWT==2.8.0
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
